librbd: allow querying if an image was thin-formatted using ImageCtx
This commit adds is_formatted_clone to ImageCtx to allow checking
if an image is a clone which was encryption formatted.
It also adds a utility function to check if an ancestor of the image has this property.
This will be useful for blocking certain APIs which are not compatible with such images,
e.g. flattening such image with encryption not loaded.

Signed-off-by: Or Ozeri <oro@il.ibm.com>