librbd/crypto: improve crypto objects lifecycle management
CryptoInterface was being a reference counted for lifecycle management.
In this commit, we remove the ref-counting, and instead have it owned
by EncryptionFormat, which in turn is owned by ImageCtx.
This commit also changes crypto::LoadRequest to actually
load new encryption instances to parent images, instead of blind-copying
the encryption instances from the child image.

Signed-off-by: Or Ozeri <oro@il.ibm.com>