librbd/crypto: switch LUKS magic on formatted clones
This commit switches the luks magic from LUKS to RBDL
when formatting a cloned image, so that if the
user tries to load the image using external LUKS parser
(such as dm-crypt) it will fail.
Note that formatted clones are not valid LUKS format,
and thus we aim to fail LUKS parsers.

Signed-off-by: Or Ozeri <oro@il.ibm.com>