librbd/crypto: add format-specific flattening operation
This commit adds a flatten operation to EncryptionFormat,
which allows defining custom format-specific flattening operation.
Specifically, when flattening a LUKS format, we replace
back the magic from RBDL to LUKS, to make the image
parse-able again by LUKS parsers, such as dm-crypt.
Furthermore, we copy the LUKS header from the parent to the child image.

Signed-off-by: Or Ozeri <oro@il.ibm.com>